Factor SpecimenColumns and SpecimenTableManager (code shared by writers and importers) out of SpecimenImporter
Migrate SpecimenWriter and SpecimenArchiveWriter to study-api
