+++ core/src/org/labkey/core/admin/AdminController.java
@@ -4694,7 +4694,7 @@ public class AdminController extends SpringActionController
             {
                 registeredFolderWriters.add(dataType);
 
-                // for each Writer also determine if their are related children Writers, if so include them also
+                // for each Writer also determine if there are related children Writers, if so include them also
                 Collection<org.labkey.api.writer.Writer> childWriters = writer.getChildren(true, true);
                 if (childWriters != null && childWriters.size() > 0)
                 {
