+++ study/api-src/org/labkey/api/study/writer/BaseStudyWriter.java
@@ -1,0 +1,10 @@
+package org.labkey.api.study.writer;
+
+import org.labkey.api.specimen.writer.SimpleStudyExportContext;
+import org.labkey.api.study.Study;
+import org.labkey.api.writer.Writer;
+
+public interface BaseStudyWriter<S extends Study, EC extends SimpleStudyExportContext> extends Writer<S, EC>
+{
+    boolean includeWithTemplate();
+}
