+++ study/src/org/labkey/study/StudyModule.java
@@ -185,7 +185,7 @@ import org.labkey.study.view.studydesign.VaccineDesignWebpartFactory;
 import org.labkey.study.writer.DatasetDataWriter;
 import org.labkey.study.writer.DefaultStudyDesignWriter;
 import org.labkey.study.writer.MissingValueWriterFactory;
-import org.labkey.study.writer.SpecimenWriter;
+import org.labkey.api.specimen.writer.SpecimenWriter;
 import org.labkey.study.writer.StudySerializationRegistryImpl;
 import org.labkey.study.writer.StudyWriterFactory;
