+++ study/src/org/labkey/study/writer/StudyWriter.java
@@ -17,6 +17,9 @@ package org.labkey.study.writer;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
+import org.labkey.api.specimen.writer.SimpleStudyExportContext;
+import org.labkey.api.study.Study;
+import org.labkey.api.study.writer.BaseStudyWriter;
 import org.labkey.api.writer.PrintWriters;
 import org.labkey.api.writer.VirtualFile;
 import org.labkey.api.writer.Writer;
@@ -48,30 +51,38 @@ public class StudyWriter implements Writer<StudyImpl, StudyExportContext>
 
         Set<String> dataTypes = ctx.getDataTypes();
 
-        // Hack for now to allow selection of CRF vs. Assay datasets.  TODO: More flexible export UI definition mechanism
-        boolean exportDatasets = dataTypes.contains(StudyArchiveDataTypes.ASSAY_DATASETS) || dataTypes.contains(StudyArchiveDataTypes.CRF_DATASETS);
-
-        // Call all the writers defined in the study module.
-        for (Writer<StudyImpl, StudyExportContext> writer : StudySerializationRegistryImpl.get().getInternalStudyWriters())
+        try
         {
-            try
+            for (BaseStudyWriter<Study, SimpleStudyExportContext> writer : StudySerializationRegistryImpl.get().getRegisteredStudyWriters())
+            {
+                String text = writer.getDataType();
+
+                if (null == text || dataTypes.contains(text))
+                    writer.write(study, ctx, vf);
+            }
+
+            // Hack for now to allow selection of CRF vs. Assay datasets.  TODO: More flexible export UI definition mechanism
+            boolean exportDatasets = dataTypes.contains(StudyArchiveDataTypes.ASSAY_DATASETS) || dataTypes.contains(StudyArchiveDataTypes.CRF_DATASETS);
+
+            // Call all the writers defined in the study module.
+            for (Writer<StudyImpl, StudyExportContext> writer : StudySerializationRegistryImpl.get().getInternalStudyWriters())
             {
                 String text = writer.getDataType();
 
                 if (null == text || dataTypes.contains(text) || (exportDatasets && text.endsWith("Datasets")))
                     writer.write(study, ctx, vf);
             }
-            catch (Exception e)
+        }
+        catch (Exception e)
+        {
+            try (OutputStream out = vf.getOutputStream("error.log"))
             {
-                try (OutputStream out = vf.getOutputStream("error.log"))
-                {
-                    // Try to get some error information to the client instead of creating a completely invalid archive
-                    PrintWriter errorWriter = PrintWriters.getPrintWriter(out);
-                    e.printStackTrace(errorWriter);
-                    errorWriter.flush();
-                }
-                throw e;
+                // Try to get some error information to the client instead of creating a completely invalid archive
+                PrintWriter errorWriter = PrintWriters.getPrintWriter(out);
+                e.printStackTrace(errorWriter);
+                errorWriter.flush();
             }
+            throw e;
         }
 
         LOG.info("Done exporting study to " + vf.getLocation());
