+++ study/src/org/labkey/study/writer/StudyWriterFactory.java
@@ -22,6 +22,9 @@ import org.labkey.api.admin.FolderWriter;
 import org.labkey.api.admin.FolderWriterFactory;
 import org.labkey.api.admin.ImportContext;
 import org.labkey.api.data.Container;
+import org.labkey.api.specimen.writer.SimpleStudyExportContext;
+import org.labkey.api.study.Study;
+import org.labkey.api.study.writer.BaseStudyWriter;
 import org.labkey.api.writer.VirtualFile;
 import org.labkey.api.writer.Writer;
 import org.labkey.folder.xml.FolderDocument;
@@ -31,7 +34,6 @@ import org.labkey.study.model.StudyManager;
 
 import java.util.ArrayList;
 import java.util.Collection;
-import java.util.LinkedList;
 import java.util.List;
 
 
@@ -118,8 +120,12 @@ public class StudyWriterFactory implements FolderWriterFactory
         public Collection<Writer> getChildren(boolean sort, boolean forTemplate)
         {
             List<Writer> children = new ArrayList<>();
-            Collection<InternalStudyWriter> writers = StudySerializationRegistryImpl.get().getInternalStudyWriters();
-            for (InternalStudyWriter writer : writers)
+            for (BaseStudyWriter<Study, SimpleStudyExportContext> writer : StudySerializationRegistryImpl.get().getRegisteredStudyWriters())
+            {
+                if (!forTemplate || writer.includeWithTemplate())
+                    children.add(writer);
+            }
+            for (InternalStudyWriter writer : StudySerializationRegistryImpl.get().getInternalStudyWriters())
             {
                 if (!forTemplate || writer.includeWithTemplate())
                     children.add(writer);
